backend/requirements.txt
psycopg
python-dotenv
Django
djangorestframework
django-filter
pillow
djangorestframework-simplejwt
django-cors-headers